requirements.txt
requests
bs4
PySocks